Consolidate change role and select role pages to one component
