+++ .vscode/eps-prescription-tracker-ui.code-workspace
@@ -65,6 +65,7 @@
         "Canonicaliser",
         "canonicalization",
         "Canonicalized",
+        "changerole",
         "Codeable",
         "codeinline",
         "codesystem",
@@ -85,6 +86,7 @@
         "Helpdesk",
         "homecare",
         "HSCN",
+        "jwks",
         "kohler",
         "liter",
         "mermade",
@@ -97,6 +99,7 @@
         "nhslogin",
         "NOSONAR",
         "OIDC",
+        "oidcjwks",
         "openapi",
         "orta",
         "Orthoptist",
@@ -131,9 +134,7 @@
         "venv",
         "versionable",
         "whens",
-        "yourselectedrole",
-        "jwks",
-        "oidcjwks"
+        "yourselectedrole"
       ],
       "cSpell.dictionaries": ["en-GB"],
       "cSpell.ignorePaths": ["package-lock.json", "node_modules", ".vscode"],
