+++ packages/cpt-ui/components/EpsRoleSelectionPage.tsx
@@ -1,0 +1,312 @@
+'use client'
+import React, {useState, useEffect, useContext, useCallback} from "react"
+import {useRouter} from 'next/navigation'
+import {Container, Col, Row, Details, Table, ErrorSummary, Button, InsetText} from "nhsuk-react-components"
+
+import {AuthContext} from "@/context/AuthProvider"
+import {useAccess} from '@/context/AccessProvider'
+
+import EpsCard, { EpsCardProps } from "@/components/EpsCard"
+import EpsSpinner from "@/components/EpsSpinner";
+
+import { RoleDetails, TrackerUserInfo } from "@/types/TrackerUserInfoTypes"
+
+// Extends the EpsCardProps to include a unique identifier
+export type RolesWithAccessProps = EpsCardProps & {
+    uuid: string
+}
+
+export type RolesWithoutAccessProps = {
+    uuid: string
+    orgName: string
+    odsCode: string
+    roleName: string
+}
+
+const trackerUserInfoEndpoint = "/api/tracker-user-info"
+
+interface RoleSelectionPageProps {
+    // contentText is where we pass in all the strings used on this page
+    contentText: {
+        title: string
+        caption: string
+        titleNoAccess: string
+        captionNoAccess: string
+        insetText: {
+            visuallyHidden: string
+            message: string
+        }
+        confirmButton: {
+            link: string
+            text: string
+        }
+        alternativeMessage: string
+        organisation: string
+        role: string
+        roles_without_access_table_title: string
+        noOrgName: string
+        rolesWithoutAccessHeader: string
+        noODSCode: string
+        noRoleName: string
+        noAddress: string
+        errorDuringRoleSelection: string
+    }
+}
+
+export default function RoleSelectionPage({ contentText }: RoleSelectionPageProps) {
+    // Destructure strings from the contentText prop
+    const {
+        title,
+        caption,
+        titleNoAccess,
+        captionNoAccess,
+        insetText,
+        confirmButton,
+        alternativeMessage,
+        organisation,
+        role,
+        roles_without_access_table_title,
+        noOrgName,
+        rolesWithoutAccessHeader,
+        noODSCode,
+        noRoleName,
+        noAddress,
+        errorDuringRoleSelection
+    } = contentText
+
+    const { setNoAccess, setSingleAccess } = useAccess()
+    const [loading, setLoading] = useState<boolean>(true)
+    const [error, setError] = useState<string | null>(null)
+    const [redirecting, setRedirecting] = useState<boolean>(false)
+    const [rolesWithAccess, setRolesWithAccess] = useState<RolesWithAccessProps[]>([])
+    const [rolesWithoutAccess, setRolesWithoutAccess] = useState<RolesWithoutAccessProps[]>([])
+
+    const router = useRouter()
+    const auth = useContext(AuthContext)
+
+    const fetchTrackerUserInfo = useCallback(async () => {
+        setLoading(true)
+        setError(null)
+        setRolesWithAccess([])
+        setRolesWithoutAccess([])
+
+        if (!auth?.isSignedIn || !auth) {
+            setLoading(false)
+            setError(null)
+            return
+        }
+
+        try {
+            const response = await fetch(trackerUserInfoEndpoint, {
+                headers: {
+                    Authorization: `Bearer ${auth?.idToken}`,
+                    'NHSD-Session-URID': '555254242106'
+                }
+            })
+
+            if (response.status !== 200) {
+                throw new Error(
+                    `Server did not return CPT user info, response ${response.status}`
+                )
+            }
+
+            const data = await response.json()
+
+            if (!data.userInfo) {
+                throw new Error("Server response did not contain data")
+            }
+
+            const userInfo: TrackerUserInfo = data.userInfo
+            const rolesWithAccess = userInfo.roles_with_access
+            const rolesWithoutAccess = userInfo.roles_without_access
+
+            setRolesWithAccess(
+                rolesWithAccess.map((role: RoleDetails, index: number) => ({
+                    uuid: `role_with_access_${index}`,
+                    orgName: role.org_name ? role.org_name : noOrgName,
+                    odsCode: role.org_code ? role.org_code : noODSCode,
+                    siteAddress: role.site_address ? role.site_address : noAddress,
+                    roleName: role.role_name ? role.role_name : noRoleName,
+                    link: "/yourselectedrole"
+                }))
+            )
+
+            setRolesWithoutAccess(
+                rolesWithoutAccess.map((role: RoleDetails, index: number) => ({
+                    uuid: `role_without_access_${index}`,
+                    roleName: role.role_name ? role.role_name : noRoleName,
+                    orgName: role.org_name ? role.org_name : noOrgName,
+                    odsCode: role.org_code ? role.org_code : noODSCode
+                }))
+            )
+
+            setNoAccess(rolesWithAccess.length === 0)
+            setSingleAccess(rolesWithAccess.length === 1)
+
+            // If the user has exactly one accessible role and zero roles without access,
+            // redirect them immediately
+            if (rolesWithAccess.length === 1 && rolesWithoutAccess.length === 0) {
+                setRedirecting(true)
+                router.push("/searchforaprescription")
+                return
+            }
+        } catch (err) {
+            setError("Failed to fetch CPT user info")
+            console.error("error fetching tracker user info:", err)
+        } finally {
+            setLoading(false)
+        }
+    }, [
+        auth,
+        router,
+        setNoAccess,
+        setSingleAccess,
+        noOrgName,
+        noODSCode,
+        noAddress,
+        noRoleName
+    ])
+
+    useEffect(() => {
+        if (auth?.isSignedIn === undefined) {
+            return
+        }
+
+        if (auth?.isSignedIn) {
+            fetchTrackerUserInfo()
+        }
+    }, [auth?.isSignedIn, fetchTrackerUserInfo])
+
+    useEffect(() => {
+        // Update error state if auth context has an error
+        setError(auth?.error ?? null)
+        if (auth?.error) {
+            setLoading(false)
+        }
+    }, [auth?.error])
+
+    if (redirecting) {
+        // Avoid rendering content if we’re about to redirect
+        return null
+    }
+
+    if (loading) {
+        return (
+            <main id="main-content" className="nhsuk-main-wrapper">
+                <Container>
+                    <Row>
+                        <Col width="full">
+                            <EpsSpinner />
+                        </Col>
+                    </Row>
+                </Container>
+            </main>
+        )
+    }
+
+    if (error) {
+        return (
+            <main id="main-content" className="nhsuk-main-wrapper">
+                <Container>
+                    <Row>
+                        <ErrorSummary>
+                            <ErrorSummary.Title>
+                                {errorDuringRoleSelection}
+                            </ErrorSummary.Title>
+                            <ErrorSummary.List>
+                                <ErrorSummary.Item href="PLACEHOLDER/contact/us">
+                                    {error}
+                                </ErrorSummary.Item>
+                            </ErrorSummary.List>
+                        </ErrorSummary>
+                    </Row>
+                </Container>
+            </main>
+        )
+    }
+
+    const noAccess = rolesWithAccess.length === 0
+
+    return (
+        <main id="main-content" className="nhsuk-main-wrapper">
+            <Container role="contentinfo">
+                <Row>
+                    <Col width="two-thirds">
+                        <h1 className="nhsuk-heading-xl">
+                            <span role="text" data-testid="eps_header_selectYourRole">
+                                <span className="nhsuk-title">
+                                    {noAccess ? titleNoAccess : title}
+                                </span>
+                                <span className="nhsuk-caption-l nhsuk-caption--bottom">
+                                    <span className="nhsuk-u-visually-hidden"> - </span>
+                                    {!noAccess && caption}
+                                </span>
+                            </span>
+                        </h1>
+                        
+                        {noAccess && <p>{captionNoAccess}</p>}
+
+                        {/* Inset Text + Confirm Button (Only if user has access) */}
+                        {!noAccess && (
+                            <section aria-label="Login Information">
+                                <InsetText>
+                                    <span className="nhsuk-u-visually-hidden">
+                                        {insetText.visuallyHidden}
+                                    </span>
+                                    <p>{insetText.message}</p>
+                                </InsetText>
+                                <Button href={confirmButton.link}>
+                                    {confirmButton.text}
+                                </Button>
+                                <p>{alternativeMessage}</p>
+                            </section>
+                        )}
+                    </Col>
+
+                    {/* Roles With Access Section */}
+                    {!noAccess && (
+                        <Col width="two-thirds">
+                            <div className="section">
+                                {rolesWithAccess.map((role: RolesWithAccessProps) => (
+                                    <EpsCard {...role} key={role.uuid} />
+                                ))}
+                            </div>
+                        </Col>
+                    )}
+
+                    {/* Roles Without Access Section */}
+                    <Col width="two-thirds">
+                        <h3>{rolesWithoutAccessHeader}</h3>
+                        <Details expander>
+                            <Details.Summary>
+                                {roles_without_access_table_title}
+                            </Details.Summary>
+                            <Details.Text>
+                                <Table>
+                                    <Table.Head>
+                                        <Table.Row>
+                                            <Table.Cell>{organisation}</Table.Cell>
+                                            <Table.Cell>{role}</Table.Cell>
+                                        </Table.Row>
+                                    </Table.Head>
+                                    <Table.Body>
+                                        {rolesWithoutAccess.map((roleItem) => (
+                                            <Table.Row key={roleItem.uuid}>
+                                                <Table.Cell>
+                                                    {roleItem.orgName} (ODS: {roleItem.odsCode})
+                                                </Table.Cell>
+                                                <Table.Cell>
+                                                    {roleItem.roleName}
+                                                </Table.Cell>
+                                            </Table.Row>
+                                        ))}
+                                    </Table.Body>
+                                </Table>
+                            </Details.Text>
+                        </Details>
+                    </Col>
+                </Row>
+            </Container>
+        </main>
+    )
+}
